src/Backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyJWT